% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalVal.R
\name{addAnnotation}
\alias{addAnnotation}
\title{A function to add the necessary annotation for **CAT**.}
\usage{
addAnnotation(gene_bed, gtf, genome_size)
}
\arguments{
\item{gene_bed}{The bed file for genes annotation.}

\item{gtf}{The GTF file.}

\item{genome_size}{The genome size file.}
}
\description{
A function to add the necessary annotation for **CAT**.
}
\examples{
  addAnnotation(gene_bed="./gene.bed", gtf="hg19.gtf", genome_size="hg19.chrom.size")
}
