% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalVal.R
\name{addFootprint}
\alias{addFootprint}
\title{A function to add necessary annotation for foorprinting analysis.}
\usage{
addFootprint(corrected_signal, bindetect_result)
}
\arguments{
\item{corrected_signal}{The absolute path of **ATACorrect** folder.}

\item{bindetect_result}{The absolute path of **BINDetect** folder.}
}
\description{
A function to add necessary annotation for foorprinting analysis.
}
\examples{
 addFootprint(corrected_signal = "./signal", bindetect_result = "./res/")
}
