% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{annoMergedPeaks}
\alias{annoMergedPeaks}
\title{Determining the merged peak category (Proximal/Distal/Intragenic).
Determining the merged peak category (Proximal / Distal / Intragenic).}
\usage{
annoMergedPeaks(quant_data, tss_flank, cutoff, save_path = NA, save_name = NA)
}
\arguments{
\item{quant_data}{Quantified data obtained by **quantification**.}

\item{tss_flank}{How many bp downstream of the TSS is considered proximal.}

\item{cutoff}{The distance of peak summit to TSS (bp) to determine the proximal / distal.}

\item{save_path}{Save file path.}

\item{save_name}{Save file name.}
}
\description{
Determining the merged peak category (Proximal/Distal/Intragenic).
Determining the merged peak category (Proximal / Distal / Intragenic).
}
