% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{annoPeaks}
\alias{annoPeaks}
\title{Annotating the peaks with ChIPSeeker.}
\usage{
annoPeaks(
  peak_path,
  sample,
  peak_suffix = "_peaks_unique.narrowPeak.bed",
  annoDb_name,
  plot_percent = T
)
}
\arguments{
\item{peak_path}{The path containing the peak bed files.}

\item{sample}{Sample names.}

\item{peak_suffix}{Provide a suffix for the peak file. Default "_peaks_unique.narrowPeak.bed".}

\item{annoDb_name}{An annotation package for your species.}

\item{plot_percent}{Plotting proportions rather than specific numbers. Default: TRUE.}
}
\description{
Annotating the peaks with ChIPSeeker.
}
\examples{
annoPeaks("~/H3ac/peaks/",c("S1","S2","S3","S4","S5","S6"),
                    "_peaks_unique.narrowPeak.bed","org.Hs.eg.db",plot_percent=F)

}
