% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibilityUtils.R
\name{calc_dist_matrix}
\alias{calc_dist_matrix}
\title{Calculate distance for matrix.}
\usage{
calc_dist_matrix(gene_range)
}
\arguments{
\item{gene_range}{}
}
\description{
Calculate distance for matrix.
}
