% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{calFisherPeaks}
\alias{calFisherPeaks}
\title{A function to calculate the two group peaks enrichment based on Fisher Test.}
\usage{
calFisherPeaks(peak_a, peak_b)
}
\arguments{
\item{peak_a}{The peak sets A you want to enrich with peak sets B.}

\item{peak_b}{The peak sets B you want to enrich with peak sets A.}

\item{genome_size}{The genome size file. Two columns: chr chr_length.}
}
\description{
A function to calculate the two group peaks enrichment based on Fisher Test.
}
\examples{
calFisherPeaks(peak_a = c("a1.bed", "a2.bed", "a3.bed"),
                         peak_b = c("b1.bed", "b2.bed", "b3.bed"), genome_size= "genome.txt")

}
