% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{calPairwise}
\alias{calPairwise}
\title{Calculate pairwised intersection of bed files.}
\usage{
calPairwise(
  bed_list,
  picture = T,
  method = "pie",
  color = "Blues",
  tl_col = "black"
)
}
\arguments{
\item{bed_list}{A list contains records of all bed files(Read by function get_beds()).}

\item{picture}{Reture plot object. Default: TRUE.}

\item{method}{Method of corrplot. Default: 'pie', can be "circle", "square", "ellipse", "number", "shade", "color", "pie".}

\item{color}{Color of glyphs.}

\item{tl_col}{Color of textg.}
}
\value{
A matrix/corrplot of pairwised intersection
}
\description{
The purpose of the cal_pairwise() function is to perform a pairwise
intersection calculation among multiple BED files and then display with
corrplot
}
\examples{
# Get beds using get_beds
beds <- get_beds(c('path1', 'path2', ... ))

plot <- cal_pairwise(beds)
}
