% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{calSPM}
\alias{calSPM}
\title{A function to quantify tissue-specificity as the Specificity Measure (SPM).}
\usage{
calSPM(quant_data, plot = TRUE)
}
\arguments{
\item{quant_data}{A normalized data matrix obtained by **quantification**.}

\item{plot}{Wheather to plot figure. Default is TRUE.}
}
\description{
A function to quantify tissue-specificity as the Specificity Measure (SPM).
}
\examples{
calSPM(quant_data)

}
