% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_res_ident}
\alias{change_res_ident}
\title{Change identity of the beds$rec based on intersection result.}
\usage{
change_res_ident(bed_list, id, rec)
}
\arguments{
\item{bed_list}{A list contains records of all bed files(Read by function get_beds()).}

\item{id}{The ID of bed in bed_list.}

\item{rec}{Result tibble from valr::intersection, contains (chrom, start, end, ident).}
}
\value{
A list which has been updated bed$res$ident based on intersection result.
}
\description{
Change identity of the beds$rec based on intersection result.
}
