% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintAnalysis.R
\name{compareFootprints}
\alias{compareFootprints}
\title{A function to compare the Tn5 signal around the motif in the whole genome.}
\usage{
compareFootprints(
  samples,
  motif,
  smooth_window = 5,
  flanking_length = 200,
  N_cores = 1,
  color = NA,
  save_path = NA,
  file_prefix = NA
)
}
\arguments{
\item{samples}{Samples used to compare the foorprint score.}

\item{motif}{Specific motifs ID. Note the inclusion of the suffix (_MAxxx).}

\item{smooth_window}{The size of the window used to smooth the lines.}

\item{flanking_length}{The flanking length to calculate and visualisation.}

\item{color}{The color list.}

\item{save_path}{The path to save the plot.}

\item{file_prefix}{The prefix of the PDF file.}
}
\description{
A function to compare the Tn5 signal around the motif in the whole genome.
}
\examples{
  compareFootprints(samples = c("CD8pos_T","Monocytes"),motif = "Arntl_MA0603.1",bindetect_path = "F:/CAT/example/BINDetect/",signal_path = "F:/CAT/example/ATACcorrect/")

}
