% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flower_plot}
\alias{flower_plot}
\title{Draw flower plot.}
\usage{
flower_plot(
  sample,
  value,
  labels,
  start = 90,
  a = 0.5,
  b = 2,
  r = 1.5,
  width = 10,
  height = 10,
  ellipse_col = rgb(135, 206, 235, 150, max = 255),
  circle_col = rgb(0, 162, 214, max = 255),
  circle_text_cex = 1
)
}
\arguments{
\item{sample}{A vector contains the name of each petal}

\item{value}{A vector contains the value of each petal}

\item{labels}{The label of the core}

\item{start}{Start deg, default is 90}

\item{a}{Width of the petal, default is 0.5}

\item{b}{Length of the petal, default is 2}

\item{r}{Radius of the core, default is 1.5}

\item{width}{Width of the plot, default is 10}

\item{height}{Heigth of the plot, default is 10}

\item{ellipse_col}{Color of the petal}

\item{circle_col}{Color of the core}

\item{circle_text_cex}{Text size of label around the petal, default is 1}
}
\value{
Flower plot
}
\description{
Draw flower plot.
}
\examples{
flower_plot(c("WSM419", "A321", "M1", "M2", "M22", "M58", "M102", "M161",
"KH36b", "KH36c", "KH36d", "KH53a", "KH53b"),c(519, 556, 83, 62, 415, 425,
357, 441, 22, 41, 33, 44, 43), 90, 0.5, 2, labels="core")
}
