% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessibilityUtils.R
\name{generate_windows}
\alias{generate_windows}
\title{Generate window for a given bed file.}
\usage{
generate_windows(win, genomic_coords)
}
\arguments{
\item{genomic_coords}{}
}
\description{
Generate window for a given bed file.
}
