% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBed.R
\name{get_beds}
\alias{get_beds}
\title{Read beds into list, can be arbitary dir_path.}
\usage{
get_beds(bedpaths, names = NULL, pattern = ".bed")
}
\arguments{
\item{bedpaths}{The vector contains the paths of bed files, can be dir_path.}

\item{names}{Define the names of input files.}

\item{pattern}{The suffix of the bed files. Default: '.bed'.}
}
\value{
A list which can be used for calculating intersection of each beds
}
\description{
Users are required to utilize this function in order to convert BED format
files into a list, which is an essential prerequisite for subsequent analyses
}
\examples{
get_beds(c('./Encode_bed/ccc.bed', './peak/xxx.bed','./peak/zzz.bed'), names=c('A1','A2','A3'))
}
