% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_combine_result}
\alias{get_combine_result}
\title{Get intersection result which can be used as input of venn/upset plot.}
\usage{
get_combine_result(bed_list)
}
\arguments{
\item{bed_list}{A list contains records of all bed files(Read by function get_beds()).}
}
\value{
A dataframe with columns key and value, where key represents combination of each bed and value represents the number of intervals.
}
\description{
Get intersection result which can be used as input of venn/upset plot.
}
\examples{
# Get beds using get_beds
beds <- get_beds(c('path1', 'path2', ... ))

# Use get_combine_result() to get intersection result
# The processing time may vary depending on the number of sets involved
inputs <- get_combine_result(beds)

}
