% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_intersect}
\alias{get_intersect}
\title{Update ident of bed_list based on intersection result}
\usage{
get_intersect(
  combine,
  bed_list,
  entrance = T,
  fparam = 1e-09,
  Fparam = 1e-09,
  rparam = F
)
}
\arguments{
\item{combine}{Combination of beds ident.}

\item{bed_list}{A list contains records of all bed files(Read by function get_beds()).}

\item{entrance}{Mark the entrance of recursion.}

\item{fparam}{Minimum overlap required as a fraction of A. Default is 1E-9 (i.e. 1bp).}

\item{Fparam}{Minimum overlap required as a fraction of B. Default is 1E-9 (i.e. 1bp).}

\item{rparam}{equire that the fraction of overlap be reciprocal for A and B.}
}
\value{
A list which has been updated ident
}
\description{
Update ident of bed_list based on intersection result
}
