% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_overlap}
\alias{get_overlap}
\title{Calculate the overlap interval.}
\usage{
get_overlap(inter_result)
}
\arguments{
\item{inter_result}{Result tibble from valr::bed_intersect, contains (start.x, end.x) and (start.y, end.y)}
}
\value{
A tibble contains the interval(start, end) of overlap region
}
\description{
Calculate the overlap interval.
}
\examples{
get_overlap(valr::bed_intersect(a, b))

}
