% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBed.R
\name{get_singlebed}
\alias{get_singlebed}
\title{Read beds into a single tibble.}
\usage{
get_singlebed(bedpaths, sort = TRUE, pattern = ".bed")
}
\arguments{
\item{bedpaths}{The vector contains the paths of bed files, can be dir_path.}

\item{sort}{A boolean value. Default: TRUE.}

\item{pattern}{The suffix of the bed files. default: '.bed'}
}
\value{
A tibble contains all the beds interval records
}
\description{
Read beds into a single tibble.
}
\examples{
get_singlebed(c('./Encode_bed/ccc.bed', './peak/xxx.bed','./peak/zzz.bed'))
}
