% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique}
\alias{get_unique}
\title{Get distinctive records of A.}
\usage{
get_unique(inter_result)
}
\arguments{
\item{inter_result}{Result tibble from valr::bed_intersect.}
}
\value{
A tibble contains distinctive intervals of x.
}
\description{
Get distinctive records of A.
}
\examples{
get_unique(valr::bed_intersect(a, b))

}
