% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_upset_input}
\alias{get_upset_input}
\title{Get expression input which can be used in upset plot}
\usage{
get_upset_input(source_df, names)
}
\arguments{
\item{source_df}{A dataframe contains key and value(From get_combine_result())}

\item{names}{A vector includes the names of each bed}
}
\value{
A vector can be used in UpSetR fromExpression
}
\description{
Get expression input which can be used in upset plot
}
\examples{
# Get beds using get_beds
beds <- get_beds(c('path1', 'path2', ... ))

# Use get_combine_result() to get intersection result
# The processing time may vary depending on the number of sets involved
inputs <- get_combine_result(beds)

get_upset_input(inputs, names)

}
