% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_venn_key}
\alias{get_venn_key}
\title{Get keys which can be used in venn plot.}
\usage{
get_venn_key(source_key, key_len)
}
\arguments{
\item{source_key}{The combination of number: 1,2,3,12,13,23,123}

\item{key_len}{The num of different numbers}
}
\value{
Binary keys such as 001,010,100...
}
\description{
Get keys which can be used in venn plot.
}
\examples{
get_venn_key(c(1,2,3,12,13,23,123), 3)

}
