% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{getClusterNum}
\alias{getClusterNum}
\title{A function to help you determine the cluster number before you use **plotClusterSpecificPeak**.}
\usage{
getClusterNum(specific_peak)
}
\arguments{
\item{specific_peak}{A sample specific peak file obtained by **getSpecificPeak**.}
}
\description{
A function to help you determine the cluster number before you use **plotClusterSpecificPeak**.
}
\examples{
 getClusterNum(sp_data)
}
