% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{getClusterTargetGenes}
\alias{getClusterTargetGenes}
\title{A function to find the target genes based on specific peaks.}
\usage{
getClusterTargetGenes(cluster_group, save_path = NA, file_prefix)
}
\arguments{
\item{cluster_group}{A group-cluster file obtained by **plotClusterSpecificPeak**.}

\item{save_path}{The path to save the sample specific peak pdf file.}

\item{file_prefix}{The save file prefix.}
}
\description{
A function to find the target genes based on specific peaks.
}
\examples{
 getClusterTargetGenes("F:/CAT/data/Specific_Peaks_Cluster_Group.txt",tss_file = "F:/CAT/data/nip_tss2.bed")

}
