% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoAccessible.R
\name{getCoaccessible}
\alias{getCoaccessible}
\title{Get the co-accessible peaks in genome wide.}
\usage{
getCoaccessible(peak_anno, quant_matrix, window = 5e+05, sample_num = 100)
}
\arguments{
\item{peak_anno}{The merged peaks annotation.}

\item{quant_matrix}{The normalized matrix obtained by **quantification**.}

\item{window}{The window size to shift.}
}
\description{
Get the co-accessible peaks in genome wide.
}
\examples{
  getCoaccessible(peak_anno="peak_anno.txt", quant_matrix="quant_mat.tsv", window=100000)
}
