% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{getCountSplit}
\alias{getCountSplit}
\title{A function to get the count file for **GWASNorm**.}
\usage{
getCountSplit(
  sample_list,
  cut_path,
  peak_path,
  peak_suffix = "_peaks_unique.narrowPeak.bed",
  save_file_path
)
}
\arguments{
\item{sample_list}{Input a vector containing the names of the samples.}

\item{cut_path}{The cuts directory obtained by running snakmake.}

\item{peak_path}{The cuts directory obtained by running snakmake.}

\item{peak_suffix}{The suffix of the peaks. Default: "_peaks_unique.narrowPeak.bed".}

\item{save_file_path}{The path to save files.}
}
\description{
A function to get the count file for **GWASNorm**.
}
\examples{
 getCountSplit(sample_list=c("s1","s2","s3","s4"), cut_path = "~/snakemake/cuts/", peak_path = "~/snakemake/peaks/", save_file_path = "~/count")
}
