% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintAnalysis.R
\name{getFootprintScore}
\alias{getFootprintScore}
\title{A function to get the footprint score.}
\usage{
getFootprintScore()
}
\value{
A matrix with all samples footprint score.
}
\description{
A function to get the footprint score.
}
\examples{
 getFootprintScore(samples=c("s1","s2","s3"...))

}
