% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{getGeneBed}
\alias{getGeneBed}
\title{Get the specific gene coordinates and flanks.}
\usage{
getGeneBed(gene_name, left = 10000, right = 10000)
}
\arguments{
\item{gene_name}{The gene name.}

\item{left}{Left flanking. Default: TRUE.}

\item{right}{Right flanking. Default: TRUE.}
}
\description{
Get the specific gene coordinates and flanks.
}
