% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASEnrichment.R
\name{getGER}
\alias{getGER}
\title{The function to integrate the GWAS enrichment results.}
\usage{
getGER(enrichment_result_path, return_matrix = FALSE)
}
\arguments{
\item{enrichment_result_path}{The GWAS enrichment result path you set in **GWASEnrichment**}

\item{return_matrix}{Whether to plot.}
}
\description{
The function to integrate the GWAS enrichment results.
}
\examples{
 getGER(enrichment_result_path = "./out")
}
