% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{getJasparMotifs2018}
\alias{getJasparMotifs2018}
\title{Get the motifs from JASPAR database (2018).}
\usage{
getJasparMotifs2018(species = "Homo sapiens", collection = "CORE", ...)
}
\arguments{
\item{species}{Species.}

\item{collection}{Collection type.}

\item{...}{}
}
\description{
Get the motifs from JASPAR database (2018).
}
\examples{
 getJasparMotifs2018(species = "Homo sapiens", collection = "CORE")
}
