% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintAnalysis.R
\name{getMotif2Gene}
\alias{getMotif2Gene}
\title{Convert motif ID to gene ID.}
\usage{
getMotif2Gene(motif_file, orgdb, tf_family)
}
\arguments{
\item{motif_file}{The PWM matrix file.}

\item{orgdb}{The orgdb name.}

\item{tf_family}{The TF family files. Thee columns are included:Symbol    Ensembl    Family.}
}
\description{
Convert motif ID to gene ID.
}
\examples{
  getMotif2Gene(motif_file = "jaspar.motif", orgdb = "org.Hs.eg.db", tf_family="tf.family.tsv")
}
