% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getNetwork}
\alias{getNetwork}
\title{Get the regulatory network.}
\usage{
getNetwork(
  sample,
  tf,
  peak2gene,
  p2g_cutoff = 0.5,
  motif2gene,
  targets = 1,
  level = 2
)
}
\arguments{
\item{sample}{The sample name. Only one sample support.}

\item{tf}{The motif ID. Only one motif support.}

\item{peak2gene}{The RDS file obtained by **getPeak2Gene**.}

\item{p2g_cutoff}{The absolute cutoff of peak2gene correlation. Default: 0.5.}

\item{motif2gene}{The annotation file obtained by **getMotif2Gene**.}

\item{targets}{The target only contain TF(1) or contain genes and TFs (2). Default: 1.}

\item{level}{Consider only the target genes of your TF of interest (1). Consider not only the target genes of your TF of interest, but also more downstream target genes (2). Default: 2.}
}
\description{
Get the regulatory network.
}
\examples{
 getNetwork(sample="Bulk_B", tf="Atf1_MA0604.1", peak2gene="All_Peak2Gene_links.rds", p2g_cutoff = 0.5, motif2gene = m2g)
}
