% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak2Gene.R
\name{getPeak2Gene}
\alias{getPeak2Gene}
\title{Calculate the ATAC-to-RNA / Peak-to-Gene links.}
\usage{
getPeak2Gene(
  atac_matrix,
  rna_matrix,
  peak_annotation,
  max_distance = 20000,
  N_permutation = 10000,
  save_path = NA
)
}
\arguments{
\item{atac_matrix}{The normalized ATAC-seq data frame obtained by **quantification**.}

\item{rna_matrix}{The normalized RNA-seq data frame. Note: The RNA and ATAC matrix have the same column names and orders.}

\item{peak_annotation}{Peak annotation obtained by **plotPDI**. The file suffix: _Proximal_Distal_Intragenic_Annotation.tsv}

\item{max_distance}{The max distance (bp) around the gene to calculate the correlations.
If you set 20000, it will find the peak-to-gene within 20kb upstream and downstream of the gene. Default: 20000.}

\item{N_permutation}{Simulate how many times pseudo-data to measure the P value of each peak-to-gene. Default: 10000.}

\item{save_path}{The path to save the result.}
}
\description{
Calculate the ATAC-to-RNA / Peak-to-Gene links.
}
\examples{
  getPeak2Gene(atac_matrix="~/zhutao/CAT/ATAC_norm_quant.tsv", rna_matrix="~/zhutao/CAT/RNA_norm_quant.tsv",
                         tss_file="~/nip_tss.bed", peak_annotation="~/zhutao/CAT/peak_anno.tsv", max_distance=20000, N_permutation=10000)
}
