% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{getRankMotif}
\alias{getRankMotif}
\title{Rank the motif deviation score.}
\usage{
getRankMotif(deviation_result, label_topN = 30, return_matrix = F)
}
\arguments{
\item{deviation_result}{The deviation result obtained by **chromVAR**.}

\item{label_topN}{The top N to label. Default is 30.}

\item{return_matrix}{Wheather to return ranked matrix. Default is FALSE.}
}
\description{
Rank the motif deviation score.
}
\examples{
  getRankMotif(deviation_result = res)

}
