% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP2Gene.R
\name{getSNP2Gene}
\alias{getSNP2Gene}
\title{A function help us look for which genes these SNPs might affect the expression of.}
\usage{
getSNP2Gene(snp, peak2gene, corr_cutoff = 0.4)
}
\arguments{
\item{snp}{The SNP coordinates file.}

\item{peak2gene}{Peak2Gene result obtained by **getPeak2Gene**.}

\item{corr_cutoff}{Pearson's correlation coefficient threshold for determining significance Peak2Gene links. Default: 0.4.}
}
\description{
A function help us look for which genes these SNPs might affect the expression of.
}
\examples{
 getSNP2Gene(snp="~/encode/CAT/test.snp", peak2gene="~/all_peak2gene.txt")
}
