% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{getSpecificPeak}
\alias{getSpecificPeak}
\title{A function to identify sample specific peaks.}
\usage{
getSpecificPeak(
  norm_data,
  jsd_file,
  cutoff = 0.2,
  save_path = NA,
  file_prefix = NA
)
}
\arguments{
\item{norm_data}{A normalized data matrix obtained by **quantification**.}

\item{jsd_file}{The JSD result obtained by **calJSD**.}

\item{cutoff}{The JSD cutoff to determine the sample specific peaks. Default is 0.2.}

\item{save_path}{The path to save the sample specific peak result.}

\item{file_prefix}{The save file prefix.}
}
\description{
A function to identify sample specific peaks.
}
\examples{
 getSpecificPeak(norm_data, jsd_file, cutoff=0.2)
}
