% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{getTimeGene}
\alias{getTimeGene}
\title{Get the fitting time genes.}
\usage{
getTimeGene(peak_time, peak2gene, rna_matrix, corr_cutoff = 0.4)
}
\arguments{
\item{peak_time}{The result of **getTimeATAC** with the parameter return_matrix = TRUE.}

\item{peak2gene}{The rds file obtained by **getPeak2Gene**.}

\item{rna_matrix}{The RNA-seq quantification matrix.}

\item{corr_cutoff}{The cutoff of correlation to get the reliable Peak2gene links. Default: 0.4.}
}
\description{
Get the fitting time genes.
}
\examples{
  getTimeGene(peak_time = peak_time, peak2gene = peak2gene,rna_matrix = rna_matrix,corr_cutoff = 0.4)
}
