% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{getTriads}
\alias{getTriads}
\title{A function to compare the chromatin accessibility across thress samples.}
\usage{
getTriads(
  group_name,
  quant_df,
  return_matrix = FALSE,
  point_size = 0.1,
  color = NA
)
}
\arguments{
\item{group_name}{The group names.}

\item{quant_df}{The normalized quantification matrix obtained by **quantification**.}

\item{return_matrix}{Whether to return a matrix instead of a plot. Default is FALSE.}

\item{point_size}{The point size. Default is 0.1.}

\item{color}{The color list.}
}
\description{
A function to compare the chromatin accessibility across thress samples.
}
\examples{
  getTriads(group_name=c("A","B","C"), quant_df = quant_df)
}
