% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{getTriadsTargetGenes}
\alias{getTriadsTargetGenes}
\title{A function to get the target genes of triads peak.}
\usage{
getTriadsTargetGenes(triads_data, save_path = NA, file_prefix = NA)
}
\arguments{
\item{triads_data}{The triads result obtained by **getTriads**.}

\item{save_path}{The path to save the result.}

\item{file_prefix}{The prefix of the file.}
}
\description{
A function to get the target genes of triads peak.
}
\examples{
  getTriadsTargetGenes(triads_data = res)
}
