% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASEnrichment.R
\name{GWASEnrichment}
\alias{GWASEnrichment}
\title{Do enrichment analysis with GWAS variants.}
\usage{
GWASEnrichment(euclideanNorm_file, snp, trait, output_path)
}
\arguments{
\item{euclideanNorm_file}{The euclideanNorm file obtained by **GWASNorm** function.}

\item{snp}{The SNPs list. It consists of two columns, the first being the chromosome and the second the position. Tabs between the two columns and no header.}

\item{trait}{The traits name you want to set.}

\item{output_path}{The path to save the enrichment results.}
}
\description{
Do enrichment analysis with GWAS variants.
}
\examples{
   GWASEnrichment(euclideanNorm_file = "Immune_Cell_counts_normToMax_quantileNorm_euclideanNorm.txt",
                             snp = "test.snp", trait = "test", output_path = "./")
}
