% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_result_df}
\alias{init_result_df}
\title{Init dataframe which would be used in intersection results based on bed ident.}
\usage{
init_result_df(n_combine, num_i)
}
\arguments{
\item{n_combine}{The number of samples.}

\item{num_i}{Combination number.}
}
\value{
Initiated dataframe with columns key and value.
}
\description{
Init dataframe which would be used in intersection results based on bed ident.
}
\examples{
init_result_df(4, 2)

}
