% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TracksUtils.R
\name{loadBigWig}
\alias{loadBigWig}
\alias{loadFullBigWig}
\title{loadBigWig}
\usage{
loadBigWig(bwFile = NULL, chr = NULL, start, end)

loadFullBigWig(bwFile = NULL)
}
\arguments{
\item{bwFile}{the path of bigwig files, default(NULL). bigwig files should end with ".bw" or ".bigwig" and directory should not be named "bw" or"bigwig".}
}
\value{
a data.frame

a data.frame
}
\description{
read bigwig files.

read bigwig files.
}
\author{
JunZhang
}
