% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBed.R
\name{parse_bedpath}
\alias{parse_bedpath}
\title{Parse bedpaths as file based path.}
\usage{
parse_bedpath(bedpaths, pattern = ".bed")
}
\arguments{
\item{bedpaths}{The vector contains the paths of bed files, can be dir_path.}

\item{pattern}{The suffix of the bed files. Default: '.bed'.}
}
\value{
A vector which contains parsed file based paths
}
\description{
Parse bedpaths as file based path.
}
\examples{
parse_bedpath(c('./Encode_bed', './peak/xxx.bed','./peak/zzz.bed'))
}
