% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{plotClusterSpecificPeak}
\alias{plotClusterSpecificPeak}
\title{A function to help you plot the sample specific peaks with heatmap.}
\usage{
plotClusterSpecificPeak(
  specific_peak,
  zscore_min = -2,
  zscore_max = 2,
  cluster_N = 10,
  cluster_col = FALSE,
  color = NA,
  save_path = NA,
  file_prefix = NA,
  figure_height = 14,
  figure_width = 7
)
}
\arguments{
\item{specific_peak}{A sample specific peak file obtained by **getSpecificPeak**.}

\item{zscore_min}{The maximum value of the zscore. Default: -2.}

\item{zscore_max}{The minium value of the zscore. Default: 2.}

\item{cluster_col}{Wheather to clustering columns. Default: FALSE.}

\item{color}{Providing your color.}

\item{save_path}{The path to save the sample specific peak pdf file.}

\item{file_prefix}{The save file prefix.}

\item{figure_height}{The PDF height. Default: 14.}

\item{figure_width}{The PDF width. Default: 7.}

\item{cluster_num}{The cluster number obtained by **getClusterNum**.}
}
\description{
A function to help you plot the sample specific peaks with heatmap.
}
\examples{
 plotClusterSpecificPeak(sp_peaks, cluster_num = 7)
}
