% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{plotCoTracks}
\alias{plotCoTracks}
\title{Plot genome tracks with co-accessible information.}
\usage{
plotCoTracks(
  samples_path,
  samples_suffix,
  chr,
  start,
  end,
  peaks,
  colist,
  coaccess_cutoff = 0.4,
  line_size = 0.8,
  curvature = 0.3,
  color = NA,
  back.color = TRUE
)
}
\arguments{
\item{samples_path}{The bigwig (bw) file folder.}

\item{samples_suffix}{The suffix of the samples. e.g. ".cpm.bw".}

\item{chr}{The chromosome ID.}

\item{start}{The start position.}

\item{end}{The end position.}

\item{peaks}{The merged peaks file.}

\item{colist}{The co-accessibile result obtained by **getCoaccessible**.}

\item{coaccess_cutoff}{The correlation cutoff of co-accessible.  Default: 0.4.}

\item{line_size}{The curve line with.  Default: 0.8.}

\item{curvature}{The degree of curvature. Default: 0.3.}

\item{color}{Color list to plot tracks.}

\item{back.color}{Whether to plot backgroud color.  Default: TRUE.}
}
\description{
Plot genome tracks with co-accessible information.
}
\examples{
 plotCoTracks(samples_path = "./signal", samples_suffix = ".cpm.bw", chr="chr1", start=10000, end=20000, peaks="final_res/meged.peaks.bed",colist="./coaccess/tsv").

}
