% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{A function used to plot tree clustering result.}
\usage{
plotDendrogram(
  norm_data,
  k,
  save_path = NA,
  figure_height = 8.12,
  figure_width = 8.12
)
}
\arguments{
\item{norm_data}{The normalized data obtained by **quantification**.}

\item{k}{How many clusters do you want to get.}

\item{save_path}{The path to save the result.}

\item{figure_height}{PDF height. Default: 8.27.}

\item{figure_width}{PDF width. Default: 8.27.}
}
\description{
A function used to plot tree clustering result.
}
\examples{
  plotDendrogram(norm_data = quant, k=10, save_path="~/CAT/")

}
