% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintAnalysis.R
\name{plotDiffFootprint}
\alias{plotDiffFootprint}
\title{Plot the differential footprint with Volcano, Bar or Lollipop.}
\usage{
plotDiffFootprint(
  samples,
  cutoff = NA,
  plot_type = c("Volcano", "Bar", "Lollipop"),
  scale = 5,
  save_path = NA,
  file_prefix = NA,
  figure_width = 8.21,
  figure_height = 8.21
)
}
\arguments{
\item{samples}{The two sample names to compare.}

\item{cutoff}{The cutoff of fold change. Default is NA.}

\item{plot_type}{Plot type. Three types are available: "Volcano","Bar","Lollipop".}

\item{scale}{}

\item{save_path}{The path to save PDF file.}

\item{file_prefix}{The prefix of PDF file.}

\item{figure_width}{PDF width. Default is 8.21.}

\item{figure_height}{PDF height. Default is 8.21.}
}
\description{
Plot the differential footprint with Volcano, Bar or Lollipop.
}
\examples{
  plotDiffFootprint(samples=c("NIP_YP1","NIP_YP2"), bindetect_path="/public/workspace/zhutao/encode/CAT",plot_type="Bar",cutoff=0.1)

}
