% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherTest.R
\name{plotEnrich}
\alias{plotEnrich}
\title{A function to plot the enrichment rsult based on **calFisherPeaks**.}
\usage{
plotEnrich(fisher_res, scale_size = c(2, 12))
}
\arguments{
\item{fisher_res}{The enrichment result obtained by **calFisherPeaks**.}

\item{scale_size}{The relative size between big and small dot size. default: c(2,12)}
}
\description{
A function to plot the enrichment rsult based on **calFisherPeaks**.
}
\examples{
 plotEnrich(res)

}
