% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotFlower}
\alias{plotFlower}
\title{Draw flower plot which represents the number of merged intervals.}
\usage{
plotFlower(
  bedpaths,
  names = NULL,
  pattern = ".bed",
  start = 90,
  a = 0.5,
  b = 2,
  r = 1.5,
  width = 10,
  height = 10,
  circle_text_cex = 1,
  circle_col = rgb(0, 162, 214, max = 255),
  ellipse_col = rgb(135, 206, 235, 150, max = 255)
)
}
\arguments{
\item{bedpaths}{The vector contains the paths of bed files, can be dir_path.}

\item{names}{Define the names of input files.}

\item{pattern}{The suffix of the bed files. Default: '.bed'.}

\item{start}{Start degree. Default: 90.}

\item{a}{Width of the petal. Default: 0.5.}

\item{b}{Length of the petal. Default: 2.}

\item{r}{Radius of the core. Default: 1.5.}

\item{width}{Width of the plot. Default: 10.}

\item{height}{Heigth of the plot. Default: 10.}

\item{circle_text_cex}{Text size of label around the petal. Default: 1}

\item{circle_col}{Color of the core.}

\item{ellipse_col}{Color of the petal.}
}
\value{
Ggplot object of venn
}
\description{
Uses can use this function to get the number of merged intervals via Flower plot
}
\examples{
draw_flower(c('path1','path2', ...))
}
