% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintAnalysis.R
\name{plotFootprintScoreExp}
\alias{plotFootprintScoreExp}
\title{Plot the correlation of footprint score and TF expression.}
\usage{
plotFootprintScoreExp(
  motif2gene,
  gene_exp,
  footprint_score,
  return_matrix = FALSE,
  cor_cutoff = 0.5,
  color = NA
)
}
\arguments{
\item{motif2gene}{The motif to gene annotation file obtained by **getMotif2Gene**.}

\item{gene_exp}{The normalized gene expression matrix.}

\item{footprint_score}{The footprint score file obtained by **getFootprintScore**.}

\item{return_matrix}{Whether to return matrix. Default is FALSE.}

\item{cor_cutoff}{The correlation cutoff of TF footprint score and TF expression. Default is 0.5.}

\item{color}{Color list.}
}
\description{
Plot the correlation of footprint score and TF expression.
}
\examples{
 plotFootprintScoreExp(motif2gene=m2g, gene_exp="rsem_exp.tsv", footprint_score = ftscore)
}
