% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotFragments}
\alias{plotFragments}
\title{A function to plot the insertion fragment size.}
\usage{
plotFragments(frag_path, sample, plot_type = "ridges")
}
\arguments{
\item{frag_path}{Provide a directory of frags from the snakemake flow.}

\item{sample}{Input a vector containing the names of the samples.}

\item{plot_type}{Option to draw ridges / density plot. Default: "ridges".}
}
\description{
A function to plot the insertion fragment size.
}
\examples{
 plotFragments("~/snakemake/fragments_size",sample=c("s1","s2"),plot_type="ridges")

}
