% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{plotGeneTracks}
\alias{plotGeneTracks}
\title{Plot specific gene ATAC signal with tracks.}
\usage{
plotGeneTracks(
  samples_path,
  samples_suffix,
  gene_name,
  left = 10000,
  right = 10000,
  peaks,
  color = NA,
  back.color = TRUE
)
}
\arguments{
\item{samples_path}{The bigwig (bw) file folder.}

\item{samples_suffix}{The suffix of the samples.}

\item{gene_name}{The gene name.}

\item{left}{The gene left flanking to plot. Default is 10000.}

\item{right}{The gene right flanking to plot. Default is 10000.}

\item{peaks}{The merged peaks.}

\item{color}{Color list to plot tracks.}

\item{back.color}{Whether to plot backgroud color.  Default if TRUE.}
}
\description{
Plot specific gene ATAC signal with tracks.
}
\examples{
 plotGeneTracks(samples_path = "./signal", samples_suffix = ".cpm.bw", gene_name="ENSG00000103888" , peaks="final_res/meged.peaks.bed")
}
