% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSampleDiffAnalysis.R
\name{plotGO}
\alias{plotGO}
\title{A function to plot the every cluster target genes GO enrichment.}
\usage{
plotGO(
  target,
  orgdb,
  N_top = 8,
  scale_size = c(2, 12),
  ont = "BP",
  clustering = TRUE
)
}
\arguments{
\item{target}{The target genes result obtained by **getClusterTargetGenes**.}

\item{orgdb}{The orgdb annotation package name.}

\item{N_top}{The top N GO terms showing in the plot. Default: 8.}

\item{scale_size}{The max and min value of dot size. Default: c(2,12).}

\item{ont}{The GO type to enrich. Three type are available: "BP", "CC", "MP". Default: BP.}

\item{clustering}{Whether to clutsering the GO terms. Default: TRUE.}
}
\description{
A function to plot the every cluster target genes GO enrichment.
}
\examples{
  plotGO(target = res, orgdb= org.Hs.eg.db, N_top = 8, scale_size=c(2,12), ont = "BP", clustering = TRUE)

}
