% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{plotMA}
\alias{plotMA}
\title{A fucntion to plot MA plot}
\usage{
plotMA(target)
}
\arguments{
\item{target}{The target genes obtained by **getDiffTargetGenes**}
}
\description{
A fucntion to plot MA plot
}
\examples{
plotMA(target)
}
