% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot the regulatory network.}
\usage{
plotNetwork(network, seed = 1)
}
\arguments{
\item{network}{The network result obtained by **getNetwork**.}

\item{seed}{Set seed number to keep the same result. Default: 1.}
}
\description{
Plot the regulatory network.
}
