% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak2Gene.R
\name{plotP2GHeatmap}
\alias{plotP2GHeatmap}
\title{Plot all the Peak2Gene links with heatmap.}
\usage{
plotP2GHeatmap(
  p2g_res,
  cor_cutoff,
  atac_matrix,
  rna_matrix,
  cluster_N = 4,
  raster = F,
  palATAC = NA,
  palRNA = NA,
  save_path = NA,
  fig_width = 12,
  fig_height = 12
)
}
\arguments{
\item{p2g_res}{The peak2gene result obtained by **getPeak2Gene**}

\item{cor_cutoff}{Pearson's correlation coefficient threshold for determining significance Peak2Gene links.}

\item{atac_matrix}{The normalized ATAC-seq data frame obtained by **quantification**. Same as **getPeak2Gene** parameter.}

\item{rna_matrix}{The normalized RNA-seq data frame. Same as **getPeak2Gene** parameter.}

\item{cluster_N}{Clustering rows into N modules. Default: 4.}

\item{raster}{Wheather to raster the plot. Default: FALSE.}

\item{palATAC}{The color of ATAC heatmap.}

\item{palRNA}{The color of RNA heatmap.}

\item{save_path}{The path to save result.}

\item{fig_width}{The PDF width. Default: 12.}

\item{fig_height}{The PDF height. Default: 12.}
}
\description{
Plot all the Peak2Gene links with heatmap.
}
\examples{
  plotP2GHeatmap("./P2G_all_links.rds",cor_cutoff=0.4,atac_matrix="./ATAC_norm_quant.tsv",rna_matrix="./RNA_norm_quant.tsv",save_path="./")

}
