% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak2Gene.R
\name{plotP2GTracks}
\alias{plotP2GTracks}
\title{Plot specific gene with tracks and peak2gene linkages.}
\usage{
plotP2GTracks(
  samples_path,
  samples_suffix,
  gene_name,
  left = 10000,
  right = 10000,
  peaks,
  peak2gene,
  cor_cutoff = 0.4,
  line_size = 0.8,
  curvature = 0.3,
  color = NA,
  back.color = TRUE
)
}
\arguments{
\item{samples_path}{The bigwig (bw) file folder.}

\item{samples_suffix}{The suffix of the samples. Like the ".cpm.bw"}

\item{gene_name}{The gene name.}

\item{left}{The gene left flanking to plot. Default: 10000.}

\item{right}{The gene right flanking to plot. Default: 10000.}

\item{peaks}{The merged peaks.}

\item{peak2gene}{The RDS file obtained by **getPeak2Gene**.}

\item{cor_cutoff}{The correlation cutoff of peak-to-gene links. Default: 0.4.}

\item{line_size}{The curve line with.  Default: 0.8.}

\item{curvature}{The degree of curvature. Default: 0.3.}

\item{color}{Color list to plot tracks.}

\item{back.colorWhether}{to plot backgroud color.  Default: TRUE.}
}
\description{
Plot specific gene with tracks and peak2gene linkages.
}
\examples{
 plotP2GTracks(samples_path="./signal", samples_suffix=".cpm.bw", gene_name="ENST00000220244", peaks="./merged_peaks.bed", peak2gene="./All_Peak2Gene_links.rds")

}
