% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot the PCA plot based on quantification.}
\usage{
plotPCA(norm_data, save_path = NA, figure_height = 8.27, figure_width = 8.27)
}
\arguments{
\item{norm_data}{The normalized data obtained by **quantification**.}

\item{save_path}{The path to save the result.}

\item{figure_height}{PDF height. Default: 8.27.}

\item{figure_width}{PDF width. Default: 8.27.}
}
\description{
Plot the PCA plot based on quantification.
}
\examples{
  plotPCA(norm_data = quant, save_path="~/CAT/")

}
