% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{plotPDI}
\alias{plotPDI}
\title{Determining the peak category (Proximal/Distal/Intragenic).}
\usage{
plotPDI(
  peak_path,
  sample,
  suffix = "_peaks_unique.narrowPeak.bed",
  tss_flank,
  cutoff,
  save_path = NA,
  save_name = NA
)
}
\arguments{
\item{peak_path}{The path containing the peak bed files.}

\item{sample}{Sample names.}

\item{suffix}{Provide a suffix for the peak file. Default: "_peaks_unique.narrowPeak.bed".}

\item{tss_flank}{How many bp downstream of the TSS is considered proximal.}

\item{cutoff}{The distance of peak summit to TSS (bp) to determine the proximal / distal.}

\item{save_path}{Save file path.}

\item{save_name}{Save file name.}
}
\description{
Determining the peak category (Proximal/Distal/Intragenic).
}
\examples{
plotPDI("F:/CAT/example/peaks/", c("Bulk_B", "Mem_B","Naive_B"),
                  "_peaks_unique.narrowPeak.bed",
                  "F:/CAT/example/hg19_gene_standard.bed", cutoff=3000, tss_flank=1000)


}
