% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak2Gene.R
\name{plotPeakGene}
\alias{plotPeakGene}
\title{Plot one gene / one peak scatter plot.}
\usage{
plotPeakGene(peak, gene, atac_matrix, rna_matrix, color = NA, legend = TRUE)
}
\arguments{
\item{peak}{The peak coordinate you want to plot.}

\item{gene}{The gene ID you want to plot.}

\item{atac_matrix}{The normalized ATAC-seq data frame obtained by **quantification**. Same as **getPeak2Gene** parameter.}

\item{rna_matrix}{The normalized RNA-seq dataframe. Same as **getPeak2Gene** parameter.}

\item{color}{The color to plot.}

\item{legend}{Wheather to plot legend. default: TRUE.}
}
\description{
Plot one gene / one peak scatter plot.
}
\examples{
  plotPeakGene("1:39081-39341","gene","ATAC_norm_quant.tsv","RNA_norm_quant.tsv",legend=F)
}
