% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{plotSummitDis}
\alias{plotSummitDis}
\title{Plot the distance of peak summit to TSS.}
\usage{
plotSummitDis(peak_path, sample, suffix)
}
\arguments{
\item{peak_path}{The path containing the peak bed files.}

\item{sample}{Sample names.}

\item{suffix}{Provide a suffix for the peak file.}
}
\description{
Plot the distance of peak summit to TSS.
}
\examples{
  plotSummitDis("~/peaks", c("Control1","Control2","Control3","Control4","Control5","IDD1","IDD2","IDD3","IDD4","IDD5"),
                          ".narrowPeak.bed")

}
