% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{plotTimeAll}
\alias{plotTimeAll}
\title{Plot the fitting time ATAC and RNA.}
\usage{
plotTimeAll(
  norm_data,
  peak2gene,
  rna_matrix,
  corr_cutoff = 0.4,
  label = NA,
  ATAC_Palette = NA,
  RNA_Palette = NA
)
}
\arguments{
\item{norm_data}{Input the normalized dynamic matrix.}

\item{peak2gene}{The peak2gene RDS file obtained by **getPeak2Gene**.}

\item{rna_matrix}{The RNA expression matrix.}

\item{corr_cutoff}{The cutoff of correlation to get the reliable Peak2gene links. Default: 0.4.}

\item{label}{The label annotation file.}

\item{ATAC_Palette}{Color for ATAC heatmap.}

\item{RNA_Palette}{Color for RNA heatmap.}
}
\description{
Plot the fitting time ATAC and RNA.
}
\examples{
 plotTimeAll(norm_data = "dynamic_ocrs.tsv", peak2gene = "F:/CAT/example/All_Peak2Gene_links.rds" , rna_matrix = "RSEM_matrix.tsv")
}
