% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{plotTimeGene}
\alias{plotTimeGene}
\title{Plot the specific gene expression along the time.}
\usage{
plotTimeGene(peak_time, peak2gene, rna_matrix, gene, corr_cutoff = 0.4)
}
\arguments{
\item{peak_time}{The result of **getTimeATAC** with the parameter return_matrix = TRUE.}

\item{peak2gene}{The rds file obtained by **getPeak2Gene**.}

\item{rna_matrix}{The RNA-seq quantification matrix.}

\item{gene}{The gene to plot.}

\item{corr_cutoff}{The cutoff of correlation to get the reliable Peak2gene links. Default: 0.4.}
}
\description{
Plot the specific gene expression along the time.
}
