% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{plotTracks}
\alias{plotTracks}
\title{Plot the genome signal with tracks.}
\usage{
plotTracks(
  samples_path,
  samples_suffix,
  chr,
  start,
  end,
  peaks,
  color = NA,
  back.color = TRUE
)
}
\arguments{
\item{samples_path}{The bigwig (bw) file folder.}

\item{samples_suffix}{The suffix of the samples. e.g. ".cpm.bw".}

\item{chr}{The chromosome ID.}

\item{start}{The start position.}

\item{end}{The end position.}

\item{peaks}{The merged peaks file.}

\item{color}{color list to plot tracks.}

\item{back.color}{Whether to plot backgroud color.  Default: TRUE.}
}
\description{
Plot the genome signal with tracks.
}
\examples{
 plotTracks(samples_path = "./signal", samples_suffix = ".cpm.bw", chr="chr1", start=10000, end=20000, peaks="final_res/meged.peaks.bed")

}
