% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{plotTriads}
\alias{plotTriads}
\title{A function to plot the triads result.}
\usage{
plotTriads(triads_data)
}
\arguments{
\item{triads_data}{The triads result obtained by **getTriads**.}
}
\description{
A function to plot the triads result.
}
\examples{
 plotTriads(triads_data = res)
}
