% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotTSS}
\alias{plotTSS}
\title{A function to plot the Tn5 signal around the TSSs.}
\usage{
plotTSS(tss_path, sample, split_group = FALSE)
}
\arguments{
\item{tss_path}{Provide a directory of TSS from the snakemake flow.}

\item{sample}{Input a vector containing the names of the samples.}

\item{split_group}{Wheather to split groups. Dafault: FALSE.}
}
\description{
A function to plot the Tn5 signal around the TSSs.
}
