% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVenn.R
\name{plotUpset}
\alias{plotUpset}
\title{Draw upset plot of bed intersection.}
\usage{
plotUpset(
  input_df,
  bed_list,
  decreasing = T,
  angles = 45,
  text_scale = 0.65,
  point_size = 2.8,
  line_size = 1
)
}
\arguments{
\item{input_df}{A dataframe contains key and value(From get_combain_result())}

\item{bed_list}{A list contains records of all bed files(Read by function get_beds())}

\item{decreasing}{See UpSetR for details. Boolean value, Default: TRUE.}

\item{angles}{The angle of main bar text, Default: 45.}

\item{text_scale}{The scale of the text. Default: 0.65.}

\item{point_size}{The point size. Default: 2.8.}

\item{line_size}{The line width. Default: 1.}
}
\value{
Upset plot of intersection result
}
\description{
This function facilitates the generation of Upset plot based on the intersection
results obtained from the get_beds() function and get_combine_result()
}
\examples{
# Get beds using get_beds
beds <- get_beds(c('path1', 'path2', ... ))

# Use get_combine_result() to get intersection result
# The processing time may vary depending on the number of sets involved
inputs <- get_combine_result(beds)

draw_upset(inputs, beds)
}
