% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialAnalysis.R
\name{plotVolcano}
\alias{plotVolcano}
\title{A function to plot the differential result.}
\usage{
plotVolcano(
  diff_data,
  save_path = NA,
  file_prefix,
  figure_height = 8.12,
  figure_width = 8.12
)
}
\arguments{
\item{diff_data}{The differential data obtained by **getDiffPeak**.}

\item{save_path}{The path to save the figure.}

\item{file_prefix}{The save file prefix.}

\item{figure_height}{The PDF figure height. Default: 8.12}

\item{figure_width}{The PDF figure width. Default: 8.12}
}
\description{
A function to plot the differential result.
}
\examples{
 plotVolcano(diff_peak)
}
