% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_normalization}
\alias{quantile_normalization}
\title{Making quantile normalization for matrix.}
\usage{
quantile_normalization(df)
}
\arguments{
\item{df}{Input the data frame to be quantile normalized.}
}
\description{
Making quantile normalization for matrix.
}
\examples{
quantile_normalization(df)
}
