% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowZscores}
\alias{rowZscores}
\title{Scale matrix with Z-score.}
\usage{
rowZscores(m = NULL, min = -2, max = 2, limit = FALSE)
}
\arguments{
\item{m}{the matrix need to zscore transform.}

\item{min}{the minimum value for the zscore. Default: -2.}

\item{max}{the maximum value for the zscore. Default: 2.}

\item{limit}{Whether to limit the maximum and minimum values of zscore. Default: FALSE.}
}
\value{
matrix after the zscore transform.
}
\description{
Scale matrix with Z-score.
}
\examples{
 rowZscores(m = mat, min = -2, max = 2, limit = TRUE)
}
